pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "probe-sniffer"
version = "0.2.0"
description = "WiFi probe request sniffer with device fingerprinting"
requires-python = ">=3.12"
license = {text = "GPL-3.0"}

dependencies = [
    "scapy>=2.5.0",
    "python-dotenv>=1.0.0",
    "paho-mqtt>=2.0.0",
    "pytz>=2024.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "requests>=2.32.0",
    "discord>=2.3.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
probe-sniffer = "probe_sniffer.capture.sniffer:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
